7-Zip Version 4.56
- some fixes in LZ encoder (LZMA and Deflate) code.
 size_t was replaces to ptrdiff_t.
 size_t version worked incorrectly with some compilers.

+++ C/Compress/Lz/MatchFinderMt.h
@@ -38,7 +38,7 @@ typedef UInt32 * (*Mf_Mix_Matches)(void *p, UInt32 matchMinPos, UInt32 *distance
 /* kMtCacheLineDummy must be >= size_of_CPU_cache_line */
 #define kMtCacheLineDummy 128
 
-typedef void (*Mf_GetHeads)(const Byte *buffer, size_t pos,
+typedef void (*Mf_GetHeads)(const Byte *buffer, UInt32 pos,
   UInt32 *hash, UInt32 hashMask, UInt32 *heads, UInt32 numHeads);
 
 typedef struct _CMatchFinderMt
